First pass at documenting / moving things to more logical places
