+++ packages/wallet-client/src/index.ts
@@ -1,7 +1,9 @@
+/**
+ * @module 
+ * The wallet client.
+ */
 
-export * from './interface/from-wallet';
-export * from './interface/to-wallet';
+export * from './wallet-events';
 
-export { decode } from './decode-utils';
 export * from './wallet-functions';
 export { WalletEventListener } from './wallet-event-listener';
