+++ packages/wallet/src/utils/state-utils.ts
@@ -1,9 +1,9 @@
 import { WalletState } from '../states';
 import { OWN_POSITION_RECEIVED, OPPONENT_POSITION_RECEIVED, WalletAction } from '../redux/actions';
-import { signatureFailure, signatureSuccess, validationFailure, validationSuccess, ResponseAction } from 'wallet-client/lib/interface/from-wallet';
+import { signatureFailure, signatureSuccess, validationFailure, validationSuccess, WalletEvent } from 'wallet-client';
 import { signPositionHex, validSignature } from './signing-utils';
 
-export function handleSignatureAndValidationMessages(state: WalletState, action: WalletAction): ResponseAction | undefined {
+export function handleSignatureAndValidationMessages(state: WalletState, action: WalletAction): WalletEvent | undefined {
   switch (action.type) {
     case OWN_POSITION_RECEIVED:
       if (state.stage !== "RUNNING") {
